refactor: :bug: improve folder structure & file naming
